% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dearseq.R
\name{summary.dearseq}
\alias{summary.dearseq}
\alias{print.summary.dearseq}
\title{Summary method for dearseq objects}
\usage{
\method{summary}{dearseq}(object, signif_threshold = 0.05, ...)

\method{print}{summary.dearseq}(x, ...)
}
\arguments{
\item{signif_threshold}{a value between \code{0} and {1} specifying the 
nominal significance threshold. Default is \code{0.05}.}

\item{...}{further arguments}

\item{x}{an object of class '\code{summary.dearseq}'.}
}
\value{
a \code{list}
}
\description{
Summary method for dearseq objects
}
\author{
Boris Hejblum
}
