% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghettiPlot_1GS.R
\name{spaghettiPlot_1GS}
\alias{spaghettiPlot_1GS}
\title{Spaghetti plot for Specific Gene Set}
\usage{
spaghettiPlot_1GS(
  gs_index,
  gmt,
  expr_mat,
  design,
  var_time,
  var_patient,
  var_group = NULL,
  var_subgroup = NULL,
  plotChoice = c("Patient medians", "Individual")
)
}
\arguments{
\item{gs_index}{index of the specific gene set in \code{gmt}.}

\item{gmt}{a \code{list} of element : geneset, geneset.name and geneset.description.}

\item{expr_mat}{a data frame with numerics of size \code{G x n} contraining the raw RNA-seq counts from \code{n} samples for \code{G} genes.}

\item{design}{a data frame containing the information of each sample (SampleID).}

\item{var_time}{the \code{time} or \code{visit} variable contained in \code{design} data.}

\item{var_patient}{the patient variable contained in \code{design} data.}

\item{var_group}{a group variable in \code{design} data to divide into two facets. Default is \code{NULL}.}

\item{var_subgroup}{a subgroup variable in \code{design} data to add 2 curves on plot for each subgroup. Default is \code{NULL}.}

\item{plotChoice}{to choose which type of plot. Default is \code{c("Patient medians", "Individual")}}
}
\value{

}
\description{
Spaghetti plot for Specific Gene Set
}
\examples{
data(baduel_5gs) 

expr_norm_corr <- as.data.frame(expr_norm_corr)
#Remove "Vernalized" columns 
expr_norm_corr <- expr_norm_corr \%>\% select(!ends_with("V")) 

#Change name of sample 
colnames(design)[1] <- "SampleID"
design <- design \%>\% filter(Vernalized == FALSE)
#Modify the Population ID to obtain value of population for each replicate
design <- design \%>\% mutate(PopulationID = Population:Replicate)

dearseq::spaghettiPlot_1GS(gs_index = 3, gmt = baduel_gmt, expr_mat = expr_norm_corr, design = design, var_time = AgeWeeks, var_patient = PopulationID, var_group = PopulationKA)
}
